chore(quality): 收紧大文件与测试噪音门禁
